Move classes around
1. Move trace_reader_writer.h to be under include/rocksdb.
2. Move FileTraceReader/Writer classes to utilities/trace.
3. Added new files file_trace_reader_writer.h/cc.
4. Added a new directory 'trace' under utilities to keep any new tracers
or tracing related code. (Thought of naming it 'tracers' initially, but
then trace sounded more generic for any future additions).
