+++ util/trace_replay.h
@@ -10,17 +10,22 @@
 #include <utility>
 
 #include "rocksdb/env.h"
-#include "rocksdb/utilities/trace_reader_writer.h"
+#include "rocksdb/trace_reader_writer.h"
 
 namespace rocksdb {
 
 class ColumnFamilyHandle;
 class DB;
 class DBImpl;
-class RandomAccessFileReader;
 class Slice;
-class WritableFileWriter;
 class WriteBatch;
+
+const std::string kTraceMagic = "feedcafedeadbeef";
+const unsigned int kTraceTimestampSize = 8;
+const unsigned int kTraceTypeSize = 1;
+const unsigned int kTracePayloadLengthSize = 4;
+const unsigned int kTraceMetadataSize =
+    kTraceTimestampSize + kTraceTypeSize + kTracePayloadLengthSize;
 
 enum TraceType : char {
   kTraceBegin = 1,
@@ -44,6 +49,7 @@ struct Trace {
   }
 };
 
+// Trace RocksDB operations using a TraceWriter.
 class Tracer {
  public:
   Tracer(Env* env, std::unique_ptr<TraceWriter>&& trace_writer);
@@ -63,6 +69,7 @@ class Tracer {
   unique_ptr<TraceWriter> trace_writer_;
 };
 
+// Replay RocksDB operations from a trace.
 class Replayer {
  public:
   Replayer(DB* db, const std::vector<ColumnFamilyHandle*>& handles,
@@ -79,36 +86,6 @@ class Replayer {
   DBImpl* db_;
   std::unique_ptr<TraceReader> trace_reader_;
   std::unordered_map<uint32_t, ColumnFamilyHandle*> cf_map_;
-};
-
-class FileTraceReader : public TraceReader {
- public:
-  FileTraceReader(std::unique_ptr<RandomAccessFileReader>&& reader);
-  ~FileTraceReader();
-
-  virtual Status Read(std::string* data) override;
-  virtual Status Close() override;
-
- private:
-  unique_ptr<RandomAccessFileReader> file_reader_;
-  Slice result_;
-  size_t offset_;
-  char* const buffer_;
-
-  static const unsigned int kBufferSize;
-};
-
-class FileTraceWriter : public TraceWriter {
- public:
-  FileTraceWriter(std::unique_ptr<WritableFileWriter>&& file_writer)
-      : file_writer_(std::move(file_writer)) {}
-  ~FileTraceWriter();
-
-  virtual Status Write(const Slice& data) override;
-  virtual Status Close() override;
-
- private:
-  unique_ptr<WritableFileWriter> file_writer_;
 };
 
 }  // namespace rocksdb
